working as a core partial
